Allow for larger values in oneshot timer intervals (#784)
* Promote InstallOneshotTimer arg from int to uint64_t in native extension
* Promote InstallOneshotTimer arg from int to BigInteger in Java class
* Use FIX2LONG in lieu of FIX2INT for interval type in t_add_oneshot_timer
* Add test for oneshot timer with large future value

+++ ext/em.cpp
@@ -1151,7 +1151,7 @@ void EventMachine_t::_RunTimers()
 EventMachine_t::InstallOneshotTimer
 ***********************************/
 
-const uintptr_t EventMachine_t::InstallOneshotTimer (int milliseconds)
+const uintptr_t EventMachine_t::InstallOneshotTimer (uint64_t milliseconds)
 {
 	if (Timers.size() > MaxOutstandingTimers)
 		return false;
